Major Rework plus added the location to ads
the data gets now pushed in a different way and is now getting passed into the recyclerview.

Changed the format of the AdModel in order to save the userID in a better way on the db.

Implemented a new activity where users can see their ads. This works fine, by checking if the userID == currentUser.uid